RQ3_RQ4/requirements.txt
openai>=1.30.0
pyyaml>=6.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
tqdm>=4.66.0
python-dateutil>=2.8.0
